Social functionality and post page cleanup
Big commit here, refactor the share icons into includes and actually
make them into working buttons. Also removed much of the stuff on the
post page which existed in the old design. I realize we're taking out
the author picture, we'll probably want to tweak the design to get that
little guy back in there.